Introduce service wrapper permitting less stateful configuration
Config files for the previous wrapper could not be shared across machines, and
could not be generated independent of or prior to creation of local system
state, as they needed to be aware of filesystem UUIDs.

This wrapper permits more flexible configuration (including support for passing
all parameters on the wrapper's command line, and not needing any on-disk files
at all) while retaining support for the legacy configuration file format.