bees: Defer full initialization until cmdline parser is done
Do not touch BeesContext until really needed. As a safety, move its
initialization after cmdline parsing. Then follow the proven design of
adding the needed default value and apply it to BeesContext later.

Also, let's set a flag when initialization after cmdline parsing is done
so we know if we already want to log an exit code.

Fixes: commit 23f3e4ec42e183cdeeb38a44445adb36434a7925
Closes: https://github.com/Zygo/bees/issues/103
Signed-off-by: Kai Krakow <kai@kaishome.de>