btrsame: a dump of random updates from 2020-2021
There isn't really a coherent development plan for this tool.  It is more
of a proving ground for various bees concepts, and occasionally useful for
spot dedupe of single large files that are known to be mostly duplicate.
It's published here because occasionally a few lines of this code become
useful for explaining to people how to implement better tools.

In particular, btrsame has no concept of btrfs's requirements for extent
splitting, so it will be ineffective as a dedupe tool for files that
are similar but not entirely identical.  Not only that, btrsame may also
create a potentially large number of unreachable blocks in the process.

Change summary:

	- opens files O_RDONLY when using dedupe_file_range ioctl

	- drop the code that implements the contiguous differences
	  threshold.

	- shorten status line by pretty-printing numbers with a unit
	  suffix

	- don't throw an error exception while in the process of
	  formatting a message about the error

	- use readahead to work around dedupe ioctl behavior

	- remove random notes about now-long-dead lockup bugs

	- remove long dead commented out file reference lookup tool

Signed-off-by: Zygo Blaxell <bees@furryterror.org>